Updated ZIO and cats
Started looking into log-effect.
Strugging with getting a LogWriter without having Wrapped into a
ZIO/ZIOs Task. I believe that the idea is to allows get the logger
during execution, might need to write a wrapper to get me the
LogWriter[F] and when comming from ZIO have a function that gets us
that.

Might need to add some logic to always have the logger being provided
with the name