Finished testing actor managment on TokenGuardian
TokenGuardian now supervises actors and has protocol to manage when
actor is properly terminated to stop it. It restars the actor if any
other exception happened
Actors that are supervised by the TokenGuardian have to use the right
message to be terminated and throw the right exception to be stopped.